Adding interpreter interface and dummy implementation
The interface itself is straight forward and only encapsulates
what’s really necessary.
All of the addition work like conflict handling and persistent data
is supposed to be done outside the interpreter’s scope.

The dummy implementation is called `nop`, which does not touch the
data at all but merely emits it. The state is not changed either.